+++ src/FSharpWrap.Tool/Program.fs
@@ -7,54 +7,22 @@ open System.IO
 open FSharpWrap.Tool.Reflection
 open FSharpWrap.Tool.Generation
 
-let private help =
-    [
-        let args =
-            Arguments.all
-            |> Map.toSeq
-            |> Seq.map snd
-        let argstr f (arg: Arguments.Info) =
-            match arg.ArgValue with
-            | "" -> id
-            | value' ->
-                fun s -> sprintf "%s <%s>" s value'
-            <| f arg.ArgType
-        args
-        |> Seq.map (argstr string)
-        |> String.concat " "
-        |> sprintf "Usage: fsharpwrap %s"
-        ""
-        "Options:"
-        ""
-        let (info, len) =
-            args
-            |> Seq.mapFold
-                (fun len' arg ->
-                    let name = argstr (fun arg' -> arg'.Name) arg
-                    (name, arg.Description), max len' name.Length)
-                0
-        yield!
-            info
-            |> Seq.map (fun (name, desc) ->
-                let ws =
-                    String.replicate (len - name.Length) " "
-                sprintf "    --%s %s%s" name ws desc)
-            |> List.ofSeq
-    ]
-
 [<EntryPoint>]
 let main argv =
-    match List.ofArray argv |> Arguments.parse with
+    match List.ofArray argv |> Options.parse with
     | Ok args ->
         if args.LaunchDebugger then
             Debugger.Launch() |> ignore
 
         // TODO: Handle errors raised during reading and writing of files
-        let file = new StreamWriter(string args.OutputFile)
+        let file =
+            let name = File.fullPath args.OutputFile
+            new StreamWriter(name.Path)
         let print =
-            Reflect.paths
-                args.Assemblies
-                args.Exclude
+            let assms =
+                List.map Path.fullPath args.Assemblies
+            args.Filter
+            |> Reflect.paths assms
             |> Generate.fromAssemblies
             |> Print.genFile
         using
@@ -65,7 +33,24 @@ let main argv =
                   Write = stream.Write }
                 |> print)
         0
+    | Error ShowUsage ->
+        printfn "Usage: fsharpwrap <assembly files> [options]"
+        stdout.WriteLine()
+        printfn "Assembly Files:"
+        printfn "  The paths to all assemblies and their dependencies"
+        stdout.WriteLine()
+        printfn "Options:"
+        for name, opt in Map.toSeq Options.all do
+            printf "  --%s" name
+            let arg =
+                match opt.ArgType with
+                | Options.Switch -> ""
+                | Options.Argument name -> sprintf " <%s>" name
+                | Options.ArgumentList names ->
+                    sprintf " [%s]" names
+            printfn "%s" arg
+            printfn "    %s" opt.Description
+        0
     | Error msg ->
-        printfn "%O" msg
-        List.iter (printfn "%s") help
+        stderr.WriteLine msg
         -1