+++ src/FSharpWrap.Tool/Reflection/Context.fs
@@ -5,9 +5,10 @@ open System.Collections.Generic
 
 open FSharpWrap.Tool
 
+[<NoComparison; NoEquality>]
 type Context =
     private
-        { Excluded: Excluded
+        { Excluded: Filter
           TypeParams: Dictionary<Type, TypeParam>
           TypeRefs: Dictionary<Type, TypeRef> }