+++ src/FSharpWrap.Tool/Reflection/Reflect.fs
@@ -7,30 +7,28 @@ open System.Reflection
 open FSharpWrap.Tool
 
 let private context r ldf filter =
+    let included = Filter.assemblyIncluded filter
     using
         (new MetadataLoadContext(r))
         (fun data ->
             let ctx = Context.init filter
             ldf data
+            |> Seq.where included
             |> Seq.map
-                (fun (assm: Assembly) ->
-                    let skip =
-                        let file = Path.GetFileName assm.Location
-                        Set.exists
-                            ((=) file)
-                            ctx.Filter.AssemblyFiles
-                    if skip
-                    then None
-                    else
-                        AssemblyInfo.ofAssembly assm ctx |> Some)
-            |> Seq.choose id
+                (fun (assm: Assembly) -> AssemblyInfo.ofAssembly assm ctx)
             |> List.ofSeq)
 
 let paths (assms: seq<Path>) =
-    let paths = Seq.map string assms
+    let files =
+        Seq.collect
+            (function
+            | Directory dir ->
+                Directory.EnumerateFiles(dir.Path, "*.dll", SearchOption.AllDirectories)
+            | File file -> Seq.singleton file.Path)
+            assms
     context
-        (PathAssemblyResolver paths)
+        (PathAssemblyResolver files)
         (fun ctx ->
             Seq.map
                 ctx.LoadFromAssemblyPath
-                paths)
+                files)